+++ apiserver/controller/src/main/java/com/zipline/controller/user/UserController.java
@@ -18,13 +18,13 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.zipline.dto.user.FindUserIdRequestDTO;
-import com.zipline.dto.user.FindUserIdResponseDTO;
-import com.zipline.dto.user.LoginRequestDTO;
-import com.zipline.dto.user.SignUpRequestDTO;
-import com.zipline.dto.user.TokenResponseDTO;
-import com.zipline.dto.user.UserModifyRequestDTO;
-import com.zipline.dto.user.UserResponseDTO;
+import com.zipline.service.user.dto.request.FindUserIdRequestDTO;
+import com.zipline.service.user.dto.response.FindUserIdResponseDTO;
+import com.zipline.service.user.dto.request.LoginRequestDTO;
+import com.zipline.service.user.dto.request.SignUpRequestDTO;
+import com.zipline.service.user.dto.response.TokenResponseDTO;
+import com.zipline.service.user.dto.request.UserModifyRequestDTO;
+import com.zipline.service.user.dto.response.UserResponseDTO;
 import com.zipline.global.jwt.dto.TokenRequestDTO;
 import com.zipline.global.response.ApiResponse;
 import com.zipline.service.user.UserService;
