+++ apiserver/service/src/main/java/com/zipline/service/user/UserServiceImpl.java
@@ -15,12 +15,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.zipline.dto.user.FindUserIdRequestDTO;
-import com.zipline.dto.user.FindUserIdResponseDTO;
-import com.zipline.dto.user.LoginRequestDTO;
-import com.zipline.dto.user.SignUpRequestDTO;
-import com.zipline.dto.user.UserModifyRequestDTO;
-import com.zipline.dto.user.UserResponseDTO;
 import com.zipline.entity.survey.Survey;
 import com.zipline.entity.user.Authority;
 import com.zipline.entity.user.User;
@@ -29,9 +23,15 @@ import com.zipline.global.exception.custom.UserNotFoundException;
 import com.zipline.global.jwt.ErrorCode;
 import com.zipline.global.jwt.TokenProvider;
 import com.zipline.global.jwt.dto.TokenRequestDTO;
-import com.zipline.repository.UserRepository;
 import com.zipline.repository.survey.SurveyRepository;
+import com.zipline.repository.user.UserRepository;
 import com.zipline.service.survey.SurveyService;
+import com.zipline.service.user.dto.request.FindUserIdRequestDTO;
+import com.zipline.service.user.dto.request.LoginRequestDTO;
+import com.zipline.service.user.dto.request.SignUpRequestDTO;
+import com.zipline.service.user.dto.request.UserModifyRequestDTO;
+import com.zipline.service.user.dto.response.FindUserIdResponseDTO;
+import com.zipline.service.user.dto.response.UserResponseDTO;
 
 import io.jsonwebtoken.JwtException;
 import lombok.RequiredArgsConstructor;
