+++ apiserver/service/src/main/java/com/zipline/service/user/dto/request/SignUpRequestDTO.java
@@ -1,4 +1,4 @@
-package com.zipline.dto.user;
+package com.zipline.service.user.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
