+++ apiserver/service/src/main/java/com/zipline/service/user/dto/request/UserModifyRequestDTO.java
@@ -1,4 +1,4 @@
-package com.zipline.dto.user;
+package com.zipline.service.user.dto.request;
 
 import java.time.LocalTime;
